% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{amelieStudy}
\alias{amelieStudy}
\title{Title How significant are your Amelie annotations?}
\usage{
amelieStudy(rndfile = NULL, ameliefile = NULL, doplot = F,
  panel = "PD", ensemble = NULL, phenotype = NULL, nNull = 100)
}
\arguments{
\item{rndfile}{A file with the RDS results of a call to \code{annotateWithAmelie()}. Note that
depending on the number of random calls you wanted to make with Amelie, calling that method
with \code{getNullDistribution} set to TRUE can take some time. Besides, once you made a call for
a phenotype, there is no need to do it again for the same phenotype.}

\item{ameliefile}{A file with the RDS results of a call to \code{annotateWithAmelie()} with
\code{getNullDistribution} set to FALSE.}

\item{doplot}{If TRUE, it will generate a plot with comparison between random chance and your
predictions in regard to number of hits and quality of predictions}

\item{panel}{A string with a pretty name for the panel as we want it to appear in the plots titles}

\item{ensemble}{The result to call \code{\link{caret::ensembleLearnKFold}}}

\item{phenotype}{A list of HPO terms that match the phenotype represented by your gene panel.
You can get your list of phenotypes by browsing at http://hpo.jax.org/}

\item{nNull}{The number of random queries for Amelie}
}
\value{
A list with three elements. They are all vectors with `nNull` + 1 length. The first
`nNull` values of all the three vectors correspond to random queries, the `nNull` + 1 is the
result of the Amelie query with your predictions. The first vector shows the number of
assocations found. The second vector the mean scores for the associations. The last vector all
results for each query. It generates a plot when `doplot` is set to TRUE.
}
\description{
Amelie annotations on you gene predictions are great as a first step to a more in deep search in
the literature for papers backing up your predictions. However, Amelie can be useful as well to
calibrate how effective are your predictions, i.e. if we are doing better than random chance when
predicting gene assocations with phenotype. This method encapsulates all the stuff for you.
}
