% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{annotateWithAmelie}
\alias{annotateWithAmelie}
\title{Annotate your new predictions with papers backing up the associations with the phenotype}
\usage{
annotateWithAmelie(ensemble, phenotype = "HP:0001300",
  getNullDistribution = F, nNull = 100)
}
\arguments{
\item{ensemble}{The result to call \code{\link{caret::ensembleLearnKFold()}}}

\item{phenotype}{A list of HPO terms that match the phenotype represented by your gene panel.
You can get your list of phenotypes by browsing at http://hpo.jax.org/}

\item{getNullDistribution}{If set to TRUE, what is does is repeating \code{nNull} times the following:
randomly select a gene set from the whole genome, with same size as predictions, ask Amelie about
associations to the phenotype, store and repeat. Normally, this way of calling \code{\link{caret::annotateWithAmelie()}}
will only be done by \code{\link{caret::amelieStudy()}}}

\item{nNull}{The number of times to repeat asking Amelie for a random gene set.}
}
\value{
It will return a data frame with a row for each single association found, and columns for the panel,
the gene, the phenotype"panel", the Amelie score, the PUBMED id of the paper, its title and the jornal of
publishing. If getNullDistribution it will add another column for indexing the random gene set call
}
\description{
Amelie is a nice tool (https://amelie.stanford.edu/) to annotate your genes with papers reporting about
Mendelian associations of your gene (or related) to a given phenotype (or phenotypes). Associations found
are scored with number in [0,100], more details at the Web site. This method takes your ensemble,
which stores your predictions and give them to Amelie to look for papers associating your genes and the
HPO terms. It can work in two different ways. It can do that for you, or it can construct a null set of
genes associated with the phenotype to compare random chance with your real predictions.
}
