% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{ensembleLearn}
\alias{ensembleLearn}
\title{Title}
\usage{
ensembleLearn(genes, panel = NULL, condition = NULL,
  expID = "BootstrapEnsemble", method = "J48", fsfile = NULL,
  vars = NULL, controls = "ghosh", filter = c("DSI", "DPI",
  "ESTcount", "constitutiveexons"), seed = 12345, nboot = 200,
  tuneLength = 5, nsamps = 5, evalctrl, evaldisease, debug = F,
  usereplacement = T, evalEach = seq(from = 1, to = nboot, by = 3),
  experiment = "DRN", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
