% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{ensembleLearnKFold}
\alias{ensembleLearnKFold}
\title{Title}
\usage{
ensembleLearnKFold(panel = "Congenital_myopathy", methods = c("rpart",
  "C5.0Tree", "svmRadial", "rf", "sparseLDA", "kknn", "naive_bayes"),
  k = 5, seed = 12345, nboot = 200, tuneLength = 5, nsamps = 5,
  auto = T, maxTrials = 5, controls = "allghosh", fs = "own",
  fsThreshold = 0.6, qmeasure = "kappa")
}
\arguments{
\item{nsamps}{}
}
\description{
Title
}
