% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featselect.R
\name{featureSelection}
\alias{featureSelection}
\title{Title}
\usage{
featureSelection(genes = NULL, seed = 12345, useSMOTE = F,
  sizes = c(5, 10, 20), k = 5, trnProp = 0.9, repeats = 10,
  gacontrols = -1)
}
\arguments{
\item{gacontrols}{}
}
\description{
Title
}
