% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{featureSelectionPlot}
\alias{featureSelectionPlot}
\title{Plotting results of your feature selection}
\usage{
featureSelectionPlot(fsdata, r = 0.6)
}
\arguments{
\item{fsdata}{The results from \code{\link{caret::featureSelection}}}

\item{r}{It has the same meaning as in \code{caret::featureSelection}. Higher the value, most selective is the filter
to select which features appear in the plot.}
}
\value{
It plots a feature selection result organized into the categories to which selected features belong to.
}
\description{
The results from a call to \code{\link{caret::featureSelection}} can be plotted with this function
}
\examples{
genes <- getGenesFromPanelApp(disorder="Neurology and neurodevelopmental disorders",
  panel="Parkinson Disease and Complex Parkinsonism", color = "green")
featureSelectionPlot(featureSelection(genes, controls = "allgenome"))
}
