% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldata.R
\name{getGenesFromPanelApp}
\alias{getGenesFromPanelApp}
\title{Title}
\usage{

  getGenesFromPanelApp(disorder = "Neurology and neurodevelopmental disorders",
  panel = "Parkinson Disease and Complex Parkinsonism", color)
}
\arguments{
\item{color}{}
}
\description{
Title
}
