% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldata.R
\name{getPanelFromPanelApp}
\alias{getPanelFromPanelApp}
\title{Title}
\usage{

  getPanelFromPanelApp(disorder = "Neurology and neurodevelopmental disorders",
  panel = "Parkinson Disease and Complex Parkinsonism")
}
\arguments{
\item{panel}{}
}
\description{
Title
}
