% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featselect.R
\name{getVarsMetaDataFromFS}
\alias{getVarsMetaDataFromFS}
\title{Which direction and to what extent do your important features affect the
gene's likelihood of being disease causing?}
\usage{
getVarsMetaDataFromFS(allfsdata, r = r, panel = "Unknown")
}
\arguments{
\item{allfsdata}{list of rfe objects. Output of \code{\link{featureSelection}}.}

\item{r}{num scalar. Between 0-1. Minimum proportion of bootstraps that must
have selected that feature as important.}

\item{panel}{chr scalar. The GE panel name from which input genes were
extracted.}
}
\value{
list with all counts, signs, effect sizes and mean effect sizes of all
 important features across all bootstraps
}
\description{
\code{getVarsFromFS} extracts from the output of
\code{\link{featureSelection}} the most important features based a voting
system across bootstraps. Then for the most-voted features obtain the sign and
effect sizes of each feature.
}
\examples{
getVarsMetaDataFromFS(allfsdata, r = 0.6, panel = "Unknown")
}
