% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Title Using a PCA data projection approach to display your ML data}
\usage{
pcaPlot(fsdata, r = 0.6, ensemble, bestPCAs = F)
}
\arguments{
\item{fsdata}{The results from \code{\link{caret::featureSelection}}}

\item{r}{It has the same meaning as in \code{caret::featureSelection}. Higher the value, most selective is the filter
to select which features appear in the plot.}

\item{ensemble}{The result to call \code{\link{caret::ensembleLearnKFold}}}

\item{bestPCAs}{If TRUE, we select PCA1 and 2 on the basis of those with best p-value on the correlation with phenotype}
}
\value{
Just a plot, nothing to return
}
\description{
Title Using a PCA data projection approach to display your ML data
}
\examples{
pcaPlot(fsdata=fspd,ensemble=pdmodel,r=0.4,bestPCAs=T)
}
