% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{seedGenesVsPredictions}
\alias{seedGenesVsPredictions}
\title{Title Still under development}
\usage{
seedGenesVsPredictions(model, panel, whereToCompare = "rea")
}
\arguments{
\item{whereToCompare}{}
}
\description{
Title Still under development
}
