% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{tsneStudy}
\alias{tsneStudy}
\title{tSNE visualization of our data and study on distance of predictions to
disease genes when comparing with random genes.}
\usage{
tsneStudy(fsdata, r = 0.6, ensemble, outpath = "~/Downloads//",
  altPanel = NULL)
}
\arguments{
\item{fsdata}{The results from \code{\link{caret::featureSelection}}}

\item{r}{It has the same meaning as in \code{caret::featureSelection}. Higher the value, most selective is the filter
to select which features appear in the plot.}

\item{ensemble}{The result to call \code{\link{caret::ensembleLearnKFold}}}

\item{outpath}{Where to store the plots}

\item{altPanel}{Pretty name for the panel, to be used at the plot. Otherwise, the one appearing
at the `ensemble` will be used}
}
\value{
Nothing, just plots into files
}
\description{
It generates the plot in a progressive manner.
In a 1st pdf, it plots all protein coding genes, with same grey color. In a second color, it
plots disease genes in orange, with their names as labels. In a 3rd plot,  we shoe neighbouring
genes for the red ones. Finally, in the last plot, predictions are shown in red.
}
\examples{
tsneStudy(fsdata=fspd,ensemble=pdmodel,r=0.4,altPanel="PD")
}
