% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anntation.R
\name{umapPlot}
\alias{umapPlot}
\title{Title Using an UMAP approach to plot the ML data projection into 2 dimensions}
\usage{
umapPlot(fsdata, r = 0.6, ensemble)
}
\arguments{
\item{fsdata}{The results from \code{\link{caret::featureSelection}}}

\item{r}{It has the same meaning as in \code{caret::featureSelection}. Higher the value, most selective is the filter
to select which features appear in the plot.}

\item{ensemble}{The result to call \code{\link{caret::ensembleLearnKFold}}}
}
\value{
A UMAP scatter plot based on the two UMAP dimensions for genes when using only the features from \code{fsdata}
as filtered by the value of \code{r}
}
\description{
Title Using an UMAP approach to plot the ML data projection into 2 dimensions
}
\examples{
umapPlot(fsdata=fspd,ensemble=pdmodel,r=0.4)
}
