% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldata.R
\name{fromGenes2MLData}
\alias{fromGenes2MLData}
\title{Title}
\usage{
fromGenes2MLData(genes, addcontrols = T, which.controls = "ghosh",
  condition = NULL, vars = NULL, filter = c("DPI", "DSI", "ESTcount",
  "constitutiveexons"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
